pyproject.toml
[project]
name = "mogemma"
version = "0.1.0"
description = "Ultra-modern Python/Mojo bridge for Google Gemma 3"
requires-python = ">=3.10"
dependencies = [
    "mojo>=0.26.1",
]

[project.optional-dependencies]
embed = [
    "numpy>=1.26.0",
]
text = [
    "numpy>=1.26.0",
    "transformers>=4.38.0",
]
hub = [
    "huggingface-hub>=0.20.0",
]
telemetry = [
    "opentelemetry-api>=1.39.1",
]
vision = [
    "numpy>=1.26.0",
    "transformers>=4.38.0",
]
all = [
    "numpy>=1.26.0",
    "transformers>=4.38.0",
    "huggingface-hub>=0.20.0",
    "opentelemetry-api>=1.39.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/py/mogemma"]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["src/py/tests"]
pythonpath = ["src/py"]

[tool.ruff]
exclude = [".venv", "node_modules", "tmp", ".tmp"]
line-length = 120
src = ["src/py/mogemma", "src/py/tests"]
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 60
preview = true
skip-magic-trailing-comma = true

[tool.ruff.lint]
extend-safe-fixes = ["TC"]
fixable = ["ALL"]
ignore = [
  "A003", "B010", "B904", "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107", "D202", "D205", "D415", "DOC501", "E501", "PLW2901", "RUF012", "ISC001", "A005", "PLC0415", "FA100", "ANN401", "FBT001", "FBT002", "FBT003", "PLR0913", "ARG002", "ARG001", "CPY001", "RUF029", "COM812", "PGH003", "PLR0911", "PLR0912", "PLR0914", "PLR0915", "PLR0916", "PLR0917", "DOC201", "BLE001", "S101", "FIX002", "TD003", "TD002", "PYI036", "A002", "DOC202", "SLF001", "PLR0904", "PLR6301", "B903", "PLW0603", "PLW0108", "RUF067"
]
select = ["ALL"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["mogemma", "tests"]
split-on-trailing-comma = false

[tool.ruff.lint.per-file-ignores]
"src/py/smoke_test.py" = ["T201", "INP001", "ANN201"]

[tool.mypy]
exclude = ["tmp/", ".tmp/"]
packages = ["mogemma", "tests"]
python_version = "3.10"
strict = true

[tool.pyright]
include = ["src/py/mogemma", "src/py/tests"]
pythonVersion = "3.10"
venv = ".venv"

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pyright>=1.1.386",
    "ruff>=0.14.14",
    "pytest>=9.0.2",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=1.3.0",
]
